+++ Gemfile
@@ -4,7 +4,10 @@ source "https://rubygems.org"
 
 # gem "rails"
 group :development, :test do
-gem "robocop", require: false
-gem "rubocop-rails", require: false
-gem "debug", require:false
+  gem "robocop", require: false
+  gem "rubocop-rails", require: false
+  gem "rubocop-minitest"
+  gem "rubocop-packaging"
+  gem "debug", require: false
+  gem "rubocop-md"
 end
